Keep a pool of emergency timers
The solution for T391426 is to give each timer its own thread, but that
makes timer creation more expensive. So, keep a pool of emergency
timers which are started and stopped as necessary, don't delete the
timer when the critical section is exited.

The CriticalSection mostly existed to hold its timer object, so with
that responsibility moved to an object with a longer lifetime,
CriticalSection can be deleted.

Bug: T391426
Change-Id: Ia2f1c9314520be40cd93fbe393425176491fb073